requirements.txt
fastapi==0.89.1
python_dateutil==2.8.2
requests==2.28.1
requests_mock==1.10.0
XlsxWriter==3.0.6
